refactor: Simplify `build_microvm_from_requests`
Simplifies `build_microvm_from_requests` by  removing generic `Fn`
arguments which where not needed.
